+++ src/bert_dataset_generation/readme.md
@@ -33,11 +33,11 @@ As `i` column in content frames is persistent across timestamps I found it more 
 ### (Lack of) Usage of `i` column
 After quick exploration of `coupons big` dataset I found out that `min_i` and `max_i` are not always aligned perfectly. Additionally, I have experienced other anomalies. Due to the lack of time I decided not to develop additional heuristics to handle them and I used prefix tree approach for `coupons big` too.
 
-# Coupon labeling dataset
+# Field extraction dataset
 This directory contains a script that can be used for creating and publishing datasets for the coupon labeling task.<br/>
 Running the script:<br/>
 ```bash
-python generate_coupon_labeling_ds.py <config_path> <ds_name> <create_repo>
+python generate_field_extraction_ds.py <config_path> <ds_name> <create_repo>
 ```
 Where `config_path` is a path to the configuration file in format demonstrated by `example_config.json` file, and
 create_repo is either 'y' or 'n'; it is used to mark whetver the dataset is being pushed to an existing or new repo. <br/>
@@ -51,6 +51,8 @@ The dataset contains a list of pairs of text sequences and labels:
 * TAG_I_DISCOUNT: inside discount text
 * TAG_B_VALIDITY: begin validity text
 * TAG_I_VALIDITY: inside validity text
+* TAG_B_ACTIVATION: inside activation text
+* TAG_I_ACTIVATION: inside activation text
 The inside labels are not used in this dataset and are expected to be used after tokenization for the model.
 ## Supported dataset formats
 The script uses only the coupons csvs. Due to vast differences in the formats of those files we will probably need to create a separate function for dataset generation for each csv that we will use. Currently supported datasets:
